feat(Page): make naming consistent (#2078)
