+++ packages/picasso/src/Page/styles.ts
@@ -1,6 +1,6 @@
 import { createStyles, Theme } from '@material-ui/core/styles'
 
-import { headerHeight } from '../TopBar/styles'
+import { headerHeight } from '../PageTopBar/styles'
 
 export default ({ screens, layout }: Theme) =>
   createStyles({
