+++ packages/picasso/src/PageTopBar/__image_snapshots__/page-topbar-default-snap.png
