+++ packages/picasso/src/PageTopBar/__image_snapshots__/page-topbar-extra-header-menu-content-snap.png
