+++ packages/picasso/src/PageTopBar/__image_snapshots__/page-topbar-left-content-snap.png
