+++ packages/picasso/src/PageTopBar/__image_snapshots__/page-topbar-right-content-snap.png
