+++ packages/picasso/src/PageTopBar/__image_snapshots__/page-topbar-with-link-snap.png
