+++ packages/picasso/src/PageTopBar/__image_snapshots__/page-topbar-without-title-snap.png
