+++ packages/picasso/src/PageTopBar/index.ts
@@ -1,0 +1,6 @@
+import { OmitInternalProps } from '@toptal/picasso-shared'
+
+import { Props } from './PageTopBar'
+
+export { default } from './PageTopBar'
+export type PageTopBarProps = OmitInternalProps<Props>
