+++ packages/picasso/src/PageTopBar/story/LeftContent.example.tsx
