+++ packages/picasso/src/PageTopBar/story/WithoutTitle.example.tsx
