+++ packages/picasso/src/PageTopBarMenu/index.ts
@@ -1,6 +1,6 @@
 import { OmitInternalProps } from '@toptal/picasso-shared'
 
-import { Props } from './TopBarMenu'
+import { Props } from './PageTopBarMenu'
 
-export { default } from './TopBarMenu'
+export { default } from './PageTopBarMenu'
 export type TopBarMenuProps = OmitInternalProps<Props>
