--- packages/picasso/src/TopBar/index.ts
@@ -1,6 +1,0 @@
-import { OmitInternalProps } from '@toptal/picasso-shared'
-
-import { Props } from './TopBar'
-
-export { default } from './TopBar'
-export type TopBarProps = OmitInternalProps<Props>
